修改代码

+++ src/types.h
@@ -5,7 +5,7 @@ typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned long u32;
 
-typedef u32 bool;
+typedef unsigned long bool;
 #define true 1
 #define false 0
